refactor(gateway): extract TurnRunner class to its own module (slice 15 of #54962)
